backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
redis[hiredis]==5.2.1
httpx==0.28.1
pydantic-settings==2.7.1
python-multipart==0.0.20
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
boto3==1.37.38
